Repl improvements
Turned off the "print result of value and save it to _" feature,
since it's broken for heap allocated things (need to prevent the GC from
gc-ing thatwhen in repl).

Now when you're writing a multi-line statement, you can use CTRL+D to
exit the multiline, but you get put back into the repl instead of
quitting.